removed old solve.js and renamed better-solve.js to solve.js
