+++ routes/api/puzzle.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const router = express.Router();
-const solve = require('../../src/better-solve.js');
+const solve = require('../../src/solve.js');
 
 router.get('/test', (req, res) => res.send('solve test'));
